+++ packages/zowe-explorer/__tests__/__mocks__/vscode.ts
@@ -918,7 +918,7 @@ export namespace l10n {
             return options;
         }
         options.args?.forEach((arg: string, i: number) => {
-            options.message = options.message.replace(`{${i}}`, arg);
+            options.message = options.message.replaceAll(`{${i}}`, arg);
         });
         return options.message;
     }
@@ -1307,11 +1307,13 @@ export enum FileSystemProviderErrorCode {
 
 /**
  * A type that filesystem providers should use to signal errors.
+ * A type that will allow Text Document selection.
+ * A type that will allow positions to be used.
  *
  * This class has factory methods for common error-cases, like `FileNotFound` when
  * a file or folder doesn't exist, use them like so: `throw vscode.FileSystemError.FileNotFound(someUri);`
  */
-export const { FileSystemError } = require("jest-mock-vscode").createVSCodeMock(jest);
+export const { FileSystemError, Selection, Position } = require("jest-mock-vscode").createVSCodeMock(jest);
 
 /**
  * Namespace for dealing with the current workspace. A workspace is the representation
@@ -1723,13 +1725,11 @@ export interface Clipboard {
     writeText(value: string): Thenable<void>;
 }
 
-export class Position {}
-
 export class Range {}
 
 export class WorkspaceEdit {
     public delete(uri: Uri, range: Range) {}
-    public insert(uri: Uri, position: Position, newText: string) {}
+    public insert(uri: Uri, position: typeof Position, newText: string) {}
 }
 
 /**
@@ -1750,5 +1750,3 @@ export namespace env {
         },
     };
 }
-
-export class Selection {}