+++ packages/zowe-explorer/__tests__/__unit__/trees/dataset/ZoweDatasetNode.unit.test.ts
@@ -10,7 +10,7 @@
  */
 
 import * as vscode from "vscode";
-import { DsEntry, Gui, imperative, PdsEntry, Validation } from "@zowe/zowe-explorer-api";
+import { BaseProvider, DsEntry, Gui, imperative, PdsEntry, Validation, ZoweScheme } from "@zowe/zowe-explorer-api";
 import * as zosfiles from "@zowe/zos-files-for-zowe-sdk";
 import {
     createSessCfgFromArgs,
@@ -28,6 +28,8 @@ import { Profiles } from "../../../../src/configuration/Profiles";
 import { ZoweLogger } from "../../../../src/tools/ZoweLogger";
 import { DatasetFSProvider } from "../../../../src/trees/dataset/DatasetFSProvider";
 import { ZoweDatasetNode } from "../../../../src/trees/dataset/ZoweDatasetNode";
+import { IconUtils } from "../../../../src/icons/IconUtils";
+import { IconGenerator } from "../../../../src/icons/IconGenerator";
 
 // Missing the definition of path module, because I need the original logic for tests
 jest.mock("fs");
@@ -120,6 +122,26 @@ describe("ZoweDatasetNode Unit Tests", () => {
         expect(testNode.label).toBeDefined();
         expect(testNode.getParent()).toBeUndefined();
         expect(testNode.getSession()).toBeDefined();
+    });
+
+    it("calls setEncoding when constructing a node with encoding", () => {
+        jest.spyOn(BaseProvider.prototype, "setEncodingForFile").mockImplementationOnce(() => {});
+        const makeEmptyDsWithEncodingMock = jest.spyOn(DatasetFSProvider.instance, "makeEmptyDsWithEncoding").mockImplementationOnce(() => {});
+        const setEncodingSpy = jest.spyOn(ZoweDatasetNode.prototype, "setEncoding");
+        const testNode = new ZoweDatasetNode({
+            label: "BRTVS99",
+            collapsibleState: vscode.TreeItemCollapsibleState.None,
+            contextOverride: Constants.DS_DS_CONTEXT,
+            encoding: { kind: "binary" },
+            profile: createIProfile(),
+            parentNode: createDatasetSessionNode(createISession(), createIProfile()),
+        });
+
+        expect(testNode.label).toBe("BRTVS99");
+        expect(testNode.collapsibleState).toBe(vscode.TreeItemCollapsibleState.None);
+        expect(setEncodingSpy).toHaveBeenCalled();
+        expect(makeEmptyDsWithEncodingMock).toHaveBeenCalled();
+        setEncodingSpy.mockRestore();
     });
 
     /*************************************************************************************************************
@@ -633,9 +655,16 @@ describe("ZoweDatasetNode Unit Tests - Function node.openDs()", () => {
 });
 
 describe("ZoweDatasetNode Unit Tests - Function node.setEncoding()", () => {
-    const setEncodingForFileMock = jest.spyOn(DatasetFSProvider.instance, "setEncodingForFile").mockImplementation();
+    let setEncodingForFileMock: jest.SpyInstance;
+    let existsMock: jest.SpyInstance;
+
+    beforeAll(() => {
+        setEncodingForFileMock = jest.spyOn(DatasetFSProvider.instance, "setEncodingForFile").mockImplementation();
+        existsMock = jest.spyOn(DatasetFSProvider.instance, "exists").mockReturnValue(true);
+    });
 
     afterAll(() => {
+        existsMock.mockRestore();
         setEncodingForFileMock.mockRestore();
     });
 
@@ -749,5 +778,150 @@ describe("ZoweDatasetNode Unit Tests - Function node.setStats", () => {
         expect(pdsEntry).not.toHaveProperty("stats");
         lookupMock.mockRestore();
         createDirMock.mockRestore();
+    });
+});
+
+describe("ZoweDatasetNode Unit Tests - function datasetRecalled", () => {
+    it("changes the collapsible state", async () => {
+        const dsNode = new ZoweDatasetNode({
+            label: "MIGRATED.PDS",
+            collapsibleState: vscode.TreeItemCollapsibleState.None,
+            contextOverride: Constants.DS_MIGRATED_FILE_CONTEXT,
+            profile: createIProfile(),
+        });
+        await (dsNode as any).datasetRecalled(true);
+        expect(dsNode.collapsibleState).toBe(vscode.TreeItemCollapsibleState.Collapsed);
+    });
+
+    it("adds a resource URI", async () => {
+        const dsNode = new ZoweDatasetNode({
+            label: "MIGRATED.PDS",
+            collapsibleState: vscode.TreeItemCollapsibleState.None,
+            contextOverride: Constants.DS_MIGRATED_FILE_CONTEXT,
+            parentNode: createDatasetSessionNode(createISession(), createIProfile()),
+            profile: createIProfile(),
+        });
+        await (dsNode as any).datasetRecalled(true);
+        expect(dsNode.resourceUri).toStrictEqual(
+            vscode.Uri.from({
+                scheme: ZoweScheme.DS,
+                path: "/sestest/MIGRATED.PDS",
+            })
+        );
+    });
+
+    it("adds a command to the node - PS", async () => {
+        const dsNode = new ZoweDatasetNode({
+            label: "MIGRATED.PS",
+            collapsibleState: vscode.TreeItemCollapsibleState.None,
+            contextOverride: Constants.DS_MIGRATED_FILE_CONTEXT,
+            parentNode: createDatasetSessionNode(createISession(), createIProfile()),
+            profile: createIProfile(),
+        });
+        await (dsNode as any).datasetRecalled(false);
+        expect(dsNode.resourceUri).toStrictEqual(
+            vscode.Uri.from({
+                scheme: ZoweScheme.DS,
+                path: "/sestest/MIGRATED.PS",
+            })
+        );
+        expect(dsNode.command).toStrictEqual({ command: "vscode.open", title: "", arguments: [dsNode.resourceUri] });
+    });
+
+    it("creates a file system entry - PDS", async () => {
+        const dsNode = new ZoweDatasetNode({
+            label: "MIGRATED.PDS",
+            collapsibleState: vscode.TreeItemCollapsibleState.None,
+            contextOverride: Constants.DS_MIGRATED_FILE_CONTEXT,
+            profile: createIProfile(),
+        });
+        const createDirMock = jest.spyOn(vscode.workspace.fs, "createDirectory").mockImplementation();
+        await (dsNode as any).datasetRecalled(true);
+        expect(createDirMock).toHaveBeenCalledWith(dsNode.resourceUri);
+    });
+
+    it("creates a file system entry - PS", async () => {
+        const dsNode = new ZoweDatasetNode({
+            label: "MIGRATED.PS",
+            collapsibleState: vscode.TreeItemCollapsibleState.None,
+            contextOverride: Constants.DS_MIGRATED_FILE_CONTEXT,
+            parentNode: createDatasetSessionNode(createISession(), createIProfile()),
+            profile: createIProfile(),
+        });
+        const writeFileMock = jest.spyOn(vscode.workspace.fs, "writeFile").mockImplementation();
+        await (dsNode as any).datasetRecalled(false);
+        expect(writeFileMock).toHaveBeenCalledWith(dsNode.resourceUri, new Uint8Array());
+    });
+
+    it("updates the icon to folder - PDS", async () => {
+        const dsNode = new ZoweDatasetNode({
+            label: "MIGRATED.PDS",
+            collapsibleState: vscode.TreeItemCollapsibleState.None,
+            contextOverride: Constants.DS_MIGRATED_FILE_CONTEXT,
+            profile: createIProfile(),
+        });
+        await (dsNode as any).datasetRecalled(true);
+        expect(dsNode.iconPath).toBe(IconGenerator.getIconById(IconUtils.IconId.folder).path);
+    });
+
+    it("updates the icon to file - PS", async () => {
+        const dsNode = new ZoweDatasetNode({
+            label: "MIGRATED.PS",
+            collapsibleState: vscode.TreeItemCollapsibleState.None,
+            contextOverride: Constants.DS_MIGRATED_FILE_CONTEXT,
+            profile: createIProfile(),
+        });
+        await (dsNode as any).datasetRecalled(false);
+        expect(dsNode.iconPath).toBe(IconGenerator.getIconById(IconUtils.IconId.document).path);
+    });
+});
+
+describe("ZoweDatasetNode Unit Tests - function datasetMigrated", () => {
+    it("changes the collapsible state", () => {
+        const dsNode = new ZoweDatasetNode({
+            label: "SOME.PDS",
+            collapsibleState: vscode.TreeItemCollapsibleState.Collapsed,
+            contextOverride: Constants.DS_PDS_CONTEXT,
+            profile: createIProfile(),
+        });
+        dsNode.datasetMigrated();
+        expect(dsNode.collapsibleState).toBe(vscode.TreeItemCollapsibleState.None);
+    });
+
+    it("removes the resource URI and command", () => {
+        const dsNode = new ZoweDatasetNode({
+            label: "SOME.PDS",
+            collapsibleState: vscode.TreeItemCollapsibleState.Collapsed,
+            contextOverride: Constants.DS_PDS_CONTEXT,
+            parentNode: createDatasetSessionNode(createISession(), createIProfile()),
+            profile: createIProfile(),
+        });
+        dsNode.datasetMigrated();
+        expect(dsNode.resourceUri).toBeUndefined();
+        expect(dsNode.command).toBeUndefined();
+    });
+
+    it("removes the file system entry", () => {
+        const dsNode = new ZoweDatasetNode({
+            label: "MIGRATED.PDS",
+            collapsibleState: vscode.TreeItemCollapsibleState.None,
+            contextOverride: Constants.DS_MIGRATED_FILE_CONTEXT,
+            profile: createIProfile(),
+        });
+        const uri = dsNode.resourceUri;
+        const removeEntryMock = jest.spyOn(DatasetFSProvider.instance, "removeEntry").mockImplementation();
+        dsNode.datasetMigrated();
+        expect(removeEntryMock).toHaveBeenCalledWith(uri);
+    });
+
+    it("changes the icon to the migrated icon", () => {
+        const dsNode = new ZoweDatasetNode({
+            label: "MIGRATED.PDS",
+            collapsibleState: vscode.TreeItemCollapsibleState.None,
+            contextOverride: Constants.DS_MIGRATED_FILE_CONTEXT,
+            profile: createIProfile(),
+        });
+        dsNode.datasetMigrated();
+        expect(dsNode.iconPath).toBe(IconGenerator.getIconById(IconUtils.IconId.migrated).path);
     });
 });