+++ packages/zowe-explorer/__tests__/__unit__/utils/LoggerUtils.unit.test.ts
@@ -46,7 +46,7 @@ function createGlobalMocks(): { [key: string]: any } {
     });
     Object.defineProperty(Gui, "infoMessage", { value: jest.fn(), configurable: true });
     jest.spyOn(vscode.workspace, "getConfiguration").mockImplementationOnce(newMocks.mockGetConfiguration);
-    Object.defineProperty(ZoweLocalStorage, "storage", {
+    Object.defineProperty(ZoweLocalStorage, "globalState", {
         value: {
             get: () => ({ persistence: true, favorites: [], history: [], sessions: ["zosmf"], searchHistory: [], fileHistory: [] }),
             update: jest.fn(),