+++ packages/zowe-explorer/src/trees/dataset/DatasetInit.ts
@@ -197,6 +197,16 @@ export class DatasetInit {
                 await datasetProvider.onDidChangeConfiguration(e);
             })
         );
+        
+        context.subscriptions.push(
+            vscode.commands.registerCommand("zowe.ds.pdsSearchFor", async (node: IZoweDatasetTreeNode) => DatasetActions.search(context, node))
+        );
+
+        context.subscriptions.push(
+            vscode.commands.registerCommand("zowe.ds.filteredDataSetsSearchFor", async (node: IZoweDatasetTreeNode) =>
+                DatasetActions.search(context, node)
+            )
+        );
 
         SharedInit.initSubscribers(context, datasetProvider);
         return datasetProvider;